backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
numpy
opencv-python-headless
scikit-learn
Pillow
matplotlib
seaborn
google-generativeai
python-dotenv